refactor: 신청 / 현재 배너 사이 상태값 분기 수정
