+++ src/platform-admin/pages/bannerApplications/BannerApplications.jsx
@@ -11,9 +11,6 @@ const bannerStatusMap = {
   ALL: '전체',
   PENDING_APPROVAL: '승인 대기',
   PENDING_PAYMENT: '결제 대기',
-  PENDING_PUBLISH: '게시 대기',
-  CANCELLED: '취소됨',
-  COMPLETED: '게시 종료',
   REJECTED: '승인 거절',
 };
 
@@ -85,12 +82,6 @@ function BannerApplications() {
       case 2:
         return 'PENDING_PAYMENT';
       case 3:
-        return 'PENDING_PUBLISH';
-      case 4:
-        return 'CANCELLED';
-      case 5:
-        return 'COMPLETED';
-      case 6:
         return 'REJECTED';
       default:
         return '';
