+++ src/platform-admin/pages/bannerApplicationsDetail/BannerApplicationsDetail.jsx
@@ -8,25 +8,18 @@ import RejectReasonModal from '../../components/rejectReasonModal/RejectReasonMo
 import RejectReasonViewModal from '../../components/rejectReasonViewModal/RejectReasonViewModal';
 import PaymentSummaryModal from '../../components/paymentSummaryModal/PaymentSummaryModal';
 import PaymentDetailModal from '../../components/paymentDetailModal/PaymentDetailModal';
-import AdCancelDetailModal from '../../components/bannerCancelDetailModal/AdCancelDetailModal';
 import ToastFail from '../../../common/components/toastFail/ToastFail';
 import { fetchDetailBanner, rejectBanner, fetchRejectInfo, fetchPaymentDetail, fetchCancelDetail } from '../../../api/service/platform-admin/banner/BannerService';
 
 const statusClassMap = {
   PENDING_APPROVAL: '승인_대기',
   PENDING_PAYMENT: '결제_대기',
-  PENDING_PUBLISH: '게시_대기',
-  CANCELLED: '취소됨',
-  COMPLETED: '게시_종료',
   REJECTED: '승인_거절',
 };
 
 const statusTextMap = {
   승인_대기: '승인 대기',
   결제_대기: '결제 대기',
-  게시_대기: '게시 대기',
-  취소됨: '취소됨',
-  게시_종료: '게시 종료',
   승인_거절: '승인 거절',
 };
 
@@ -39,7 +32,6 @@ function BannerApplicationsDetail() {
   const [showPaymentModal, setShowPaymentModal] = useState(false);
   const [showPaymentDetail, setShowPaymentDetail] = useState(false);
   const [showRejectViewModal, setShowRejectViewModal] = useState(false);
-  const [showCancelDetail, setShowCancelDetail] = useState(false);
 
   const [bannerData, setBannerData] = useState(null);
   const [operatorData, setOperatorData] = useState(null);
@@ -52,7 +44,6 @@ function BannerApplicationsDetail() {
   const [failMessage, setFailMessage] = useState('');
   const [rejectReason, setRejectReason] = useState(null);
   const [paymentDetail, setPaymentDetail] = useState(null);
-  const [cancelDetail, setCancelDetail] = useState(null);
 
   const getRejectReason = async () => {
     try {
@@ -73,15 +64,6 @@ function BannerApplicationsDetail() {
     }
   }
 
-  const getCancelDetail = async () => {
-    try {
-      const res = await fetchCancelDetail(id);
-      setCancelDetail(res);
-      console.log("cancelInfo = ", res); // todo: 삭제
-    } catch (err) {
-      console.log("결제 정보를 불러오지 못했습니다 : ", err);
-    }
-  }
 
   const fetchData = async () => {
     try {
@@ -93,9 +75,6 @@ function BannerApplicationsDetail() {
       } else if (rawStatus == 'CANCELLED') {
         getPaymentDetail();
         getCancelDetail();
-      }else if (rawStatus == 'PENDING_PAYMENT' || rawStatus == 'COMPLETED' 
-        || rawStatus == 'PENDING_PUBLISH') {
-        getPaymentDetail();
       }
       setBannerData(response);
       setOperatorData({
@@ -173,21 +152,7 @@ function BannerApplicationsDetail() {
         <button className={styles.approveBtn} onClick={() => setShowRejectViewModal(true)}>거절 사유</button>
       </div>
     );
-  } else if (rawStatus === 'CANCELLED') {
-    buttonGroup = (
-      <div className={styles.buttonGroup}>
-        <button className={styles.approveBtn} onClick={() => setShowPaymentDetail(true)}>결제 내역</button>
-        <button className={styles.approveBtn} onClick={() => setShowCancelDetail(true)}>취소 내역</button>
-      </div>
-    );
-  } else if (rawStatus === 'COMPLETED' || rawStatus === 'PENDING_PAYMENT') {
-    buttonGroup = (
-      <div className={styles.buttonGroup}>
-        <button className={styles.approveBtn} onClick={() => setShowPaymentDetail(true)}>결제 내역</button>
-      </div>
-    );
   }
-
   return (
     <div className={styles.operatorContainer}>
       {/* 상단 제목 및 상태 뱃지 */}
@@ -230,18 +195,12 @@ function BannerApplicationsDetail() {
         onClose={() => setShowPaymentModal(false)}
         onSubmit={handleApproveSubmit}
       />
-
       <PaymentDetailModal
         isOpen={showPaymentDetail}
         onClose={() => setShowPaymentDetail(false)}
         paymentDetail={paymentDetail}
       />
 
-      <AdCancelDetailModal
-        isOpen={showCancelDetail}
-        onClose={() => setShowCancelDetail(false)}
-        cancelDetail = {cancelDetail}
-      />
 
       {showFailToast && <ToastFail message={failMessage} />}
     </div>
