+++ src/platform-admin/pages/bannerApplicationsDetail/BannerApplicationsDetail.module.css
@@ -66,21 +66,10 @@
   background-color: #f0ad4e;
 }
 
-.게시_대기 {
-  background-color: #5cb85c;
-}
-
-.취소됨 {
-  background-color: #d9534f;
-}
-
 .승인_거절 {
   background-color: #d9534f;
 }
 
-.게시_종료 {
-  background-color: #6c757d;
-}
 
 /* --- 버튼 그룹 --- */
 .buttonGroup {
