+++ src/platform-admin/pages/bannerCurrentDetail/BannerCurrentDetail.jsx
@@ -6,6 +6,7 @@ import BannerApplicationForm from '../../components/bannerApplicationForm/Banner
 import OperatorApplicationForm from '../../components/operatorApplicationForm/OperatorApplicationForm';
 import PaymentDetailModal from '../../components/paymentDetailModal/PaymentDetailModal';
 import SettlementDetailModal from '../../components/settlementDetailModal/SettlementDetailModal';
+import AdCancelDetailModal from '../../components/bannerCancelDetailModal/AdCancelDetailModal';
 import SettlementSummaryModal from '../../components/settlementSummaryModal/SettlementSummaryModal';
 import ToastFail from '../../../common/components/toastFail/ToastFail';
 import { fetchDetailBanner, fetchCancelInfo, fetchPaymentDetail, cancelBanner } from '../../../api/service/platform-admin/banner/BannerService';
@@ -13,11 +14,17 @@ import { fetchDetailBanner, fetchCancelInfo, fetchPaymentDetail, cancelBanner } 
 const statusClassMap = {
   PENDING_CANCEL: '취소_대기',
   PUBLISHED: '게시중',
+  PENDING_PUBLISH: '게시_대기',
+  CANCELLED: '취소됨',
+  COMPLETED: '게시_종료',
 };
 
 const statusTextMap = {
   취소_대기: '취소 대기',
   게시중: '게시중',
+  게시_대기: '게시 대기',
+  취소됨: '취소됨',
+  게시_종료: '게시 종료'
 };
 
 function BannerCurrentDetail() {
@@ -28,6 +35,7 @@ function BannerCurrentDetail() {
   const [showSettlementSummary, setShowSettlementSummary] = useState(false);
   const [showSettlementDetail, setShowSettlementDetail] = useState(false);
   const [showPaymentDetail, setShowPaymentDetail] = useState(false);
+  const [showCancelDetail, setShowCancelDetail] = useState(false);
 
   const [bannerData, setBannerData] = useState(null);
   const [operatorData, setOperatorData] = useState(null);
@@ -36,6 +44,7 @@ function BannerCurrentDetail() {
   const statusClass = statusClassMap[rawStatus];
   const statusText = statusTextMap[statusClass];
 
+  const [cancelDetail, setCancelDetail] = useState(null);
   const [cancelForm, setCancelForm] = useState(null);
   const [showFailToast, setShowFailToast] = useState(false);
   const [failMessage, setFailMessage] = useState('');
@@ -57,6 +66,16 @@ function BannerCurrentDetail() {
       console.log("fetch cancel info", res);
     } catch (err) {
       console.log("취소 정보를 불러오지 못했습니다:", err);
+    }
+  }
+
+    const getCancelDetail = async () => {
+    try {
+      const res = await fetchCancelDetail(id);
+      setCancelDetail(res);
+      console.log("cancelInfo = ", res); // todo: 삭제
+    } catch (err) {
+      console.log("결제 정보를 불러오지 못했습니다 : ", err);
     }
   }
 
@@ -82,6 +101,7 @@ function BannerCurrentDetail() {
         getPaymentDetail();
 
         if (rawStatus == 'PENDING_CANCEL') {
+          getPaymentDetail();
           fetchCancelForm();
           console.log("cancelForm : ", cancelForm);
         }
@@ -122,21 +142,30 @@ function BannerCurrentDetail() {
     buttonGroup = (
       <div className={styles.buttonGroup}>
         <button className={styles.approveBtn} onClick={() => setShowPaymentDetail(true)}>결제 내역</button>
-        <button
-          className={styles.approveBtn}
-          onClick={() => setShowSettlementSummary(true)}
-        >
-          취소 승인
-        </button>
+        <button className={styles.approveBtn} onClick={() => setShowSettlementSummary(true)}>취소 승인</button>
       </div>
     );
-  }else if (rawStatus === 'PUBLISHED'){
-        buttonGroup = (
+  } else if (rawStatus === 'PUBLISHED') {
+    buttonGroup = (
+      <div className={styles.buttonGroup}>
+        <button className={styles.approveBtn} onClick={() => setShowPaymentDetail(true)}>결제 내역</button>
+      </div>
+    );
+  } else if (rawStatus === 'CANCELLED') {
+    buttonGroup = (
+      <div className={styles.buttonGroup}>
+        <button className={styles.approveBtn} onClick={() => setShowPaymentDetail(true)}>결제 내역</button>
+        <button className={styles.approveBtn} onClick={() => setShowCancelDetail(true)}>취소 내역</button>
+      </div>
+    );
+  } else if (rawStatus === 'COMPLETED' || rawStatus === 'PENDING_PAYMENT') {
+    buttonGroup = (
       <div className={styles.buttonGroup}>
         <button className={styles.approveBtn} onClick={() => setShowPaymentDetail(true)}>결제 내역</button>
       </div>
     );
   }
+
 
   return (
     <div className={styles.operatorContainer}>
@@ -179,6 +208,13 @@ function BannerCurrentDetail() {
         isOpen={showPaymentDetail}
         onClose={() => setShowPaymentDetail(false)}
       />
+  
+      <AdCancelDetailModal
+        isOpen={showCancelDetail}
+        onClose={() => setShowCancelDetail(false)}
+        cancelDetail = {cancelDetail}
+      />
+
 
       {showFailToast && <ToastFail message={failMessage} />}
     </div>
